Refactoring blur
1. Refactoring blur, see discussion in : 7f8bb80
2. Releated issue: #330
