--- Wox.Core/Theme/ITheme.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Wox.Core.Theme
-{
-    interface ITheme
-    {
-        void ChangeTheme(string themeName);
-        List<string> LoadAvailableThemes();
-    }
-}
