+++ Wox/Languages/ru.xaml
@@ -43,7 +43,6 @@
     <system:String x:Key="resultItemFont">Шрифт результатов</system:String>
     <system:String x:Key="windowMode">Оконный режим</system:String>
     <system:String x:Key="opacity">Прозрачность</system:String>
-    <system:String x:Key="windowsAccents">Windows Accents</system:String>
     
     <!--Setting Hotkey-->
     <system:String x:Key="hotkey">Горячие клавиши</system:String>
