+++ Wox/Languages/zh-cn.xaml
@@ -43,7 +43,6 @@
     <system:String x:Key="resultItemFont">结果项字体</system:String>
     <system:String x:Key="windowMode">窗口模式</system:String>
     <system:String x:Key="opacity">透明度</system:String>
-    <system:String x:Key="windowsAccents">Windows Accents</system:String>
     
     <!--设置，热键-->
     <system:String x:Key="hotkey">热键</system:String>
