+++ Wox/Themes/Base.xaml
@@ -27,10 +27,6 @@
         <Setter Property="Stroke" Value="Blue" />
     </Style>
 
-
-    <system:Boolean x:Key="ThemeBlurEnabled">false</system:Boolean>
-    
-    
     <!-- Item Style -->
     <Style x:Key="BaseItemTitleStyle" TargetType="{x:Type TextBlock}">
         <Setter Property="Foreground" Value="#FFFFF8" />
